feat: merged CourseTakenModule with CourseModule
