+++ src/CourseModule/dto/alternative-progression-data.dto.ts
@@ -1,0 +1,3 @@
+import { TestWithoutCorrectAlternativeDTO } from './test-without-correct-alternative.dto';
+
+export class AlternativeProgressionDataDTO extends TestWithoutCorrectAlternativeDTO {}
