+++ src/CourseModule/dto/video-progression-data.dto.ts
