+++ src/CourseModule/entity/lesson.entity.ts
@@ -9,7 +9,7 @@ import {
 } from 'typeorm';
 import { Course } from './course.entity';
 import { Part } from './part.entity';
-import { CourseTaken } from '../../CourseTakenModule/entity/course.taken.entity';
+import { CourseTaken } from './course.taken.entity';
 import { Audit } from '../../CommonsModule/entity/audit.entity';
 
 @Unique(['sequenceNumber', 'course'])
