+++ src/CourseModule/entity/part.entity.ts
@@ -10,7 +10,7 @@ import {
 import { Lesson } from './lesson.entity';
 import { Test } from './test.entity';
 import { Expose } from 'class-transformer';
-import { CourseTaken } from '../../CourseTakenModule/entity/course.taken.entity';
+import { CourseTaken } from './course.taken.entity';
 import { Audit } from '../../CommonsModule/entity/audit.entity';
 
 @Unique(['sequenceNumber', 'lesson'])
