+++ src/CourseModule/entity/test.entity.ts
@@ -9,7 +9,7 @@ import {
 } from 'typeorm';
 import { Part } from './part.entity';
 import { Expose } from 'class-transformer';
-import { CourseTaken } from '../../CourseTakenModule/entity/course.taken.entity';
+import { CourseTaken } from './course.taken.entity';
 import { Audit } from '../../CommonsModule/entity/audit.entity';
 
 @Unique(['sequenceNumber', 'part'])
