+++ src/CourseModule/swagger/new-course.taken.swagger.ts
