--- src/CourseTakenModule/dto/alternative-progression-data.dto.ts
@@ -1,3 +1,0 @@
-import { TestWithoutCorrectAlternativeDTO } from '../../CourseModule/dto/test-without-correct-alternative.dto';
-
-export class AlternativeProgressionDataDTO extends TestWithoutCorrectAlternativeDTO {}
