+++ src/DashboardModule/dashboard.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { DashboardService } from './service/dashboard.service';
 import { DashboardController } from './controller/dashboard.controller';
 import { UserModule } from '../UserModule/user.module';
-import { CourseTakenModule } from '../CourseTakenModule/course.taken.module';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
+import { CourseModule } from '../CourseModule/course.module';
 
 @Module({
   imports: [
@@ -18,7 +18,7 @@ import { ConfigService } from '@nestjs/config';
       inject: [ConfigService],
     }),
     UserModule,
-    CourseTakenModule,
+    CourseModule,
   ],
   controllers: [DashboardController],
   providers: [DashboardService],
