+++ src/UserModule/controller/school.controller.ts
@@ -1,5 +1,5 @@
 import { SchoolService } from '../service/school.service';
-import { Controller, Get, UseGuards, Query } from '@nestjs/common';
+import { Controller, Get, Query, UseGuards } from '@nestjs/common';
 import { NeedRole } from '../../CommonsModule/guard/role-metadata.guard';
 import { RoleEnum } from '../../SecurityModule/enum/role.enum';
 import { RoleGuard } from '../../CommonsModule/guard/role.guard';
