+++ src/UserModule/entity/user.entity.ts
@@ -10,7 +10,7 @@ import {
 import { Role } from '../../SecurityModule/entity/role.entity';
 import { ChangePassword } from './change-password.entity';
 import { Expose } from 'class-transformer';
-import { CourseTaken } from '../../CourseTakenModule/entity/course.taken.entity';
+import { CourseTaken } from '../../CourseModule/entity/course.taken.entity';
 import { Audit } from '../../CommonsModule/entity/audit.entity';
 import { Certificate } from '../../CertificateModule/entity/certificate.entity';
 import { GenderEnum } from '../enum/gender.enum';
