+++ src/UserModule/user.module.ts
@@ -1,4 +1,4 @@
-import { forwardRef, Module, HttpModule, HttpService } from '@nestjs/common';
+import { forwardRef, HttpModule, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
