+++ src/main.ts
@@ -1,4 +1,3 @@
-import * as PubSub from 'pubsub-js';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
