+++ test/DashboardModule/dashboard.e2e-spec.ts
@@ -13,13 +13,13 @@ import { GrantTypeEnum } from '../../src/SecurityModule/enum/grant-type.enum';
 import { NewCourseDTO } from '../../src/CourseModule/dto/new-course.dto';
 import { Constants } from '../../src/CommonsModule/constants';
 import { NewUserDTO } from '../../src/UserModule/dto/new-user.dto';
-import { CourseTaken } from '../../src/CourseTakenModule/entity/course.taken.entity';
+import { CourseTaken } from '../../src/CourseModule/entity/course.taken.entity';
 import { GenderEnum } from '../../src/UserModule/enum/gender.enum';
 import { EscolarityEnum } from '../../src/UserModule/enum/escolarity.enum';
 import { UserService } from '../../src/UserModule/service/user.service';
 import { CourseService } from '../../src/CourseModule/service/course.service';
-import { CourseTakenService } from '../../src/CourseTakenModule/service/course.taken.service';
-import { CourseTakenStatusEnum } from '../../src/CourseTakenModule/enum/enum';
+import { CourseTakenService } from '../../src/CourseModule/service/course.taken.service';
+import { CourseTakenStatusEnum } from '../../src/CourseModule/enum/enum';
 import { UserProfileEnum } from '../../src/UserModule/enum/user-profile.enum';
 import { OrderEnum } from '../../src/DashboardModule/enum/order.enum';
 import { REQUEST } from '@nestjs/core';
